feat: add scan command
