+++ docs/user/commands/kyverno-json.md
@@ -13,14 +13,12 @@ kyverno-json [flags]
 ### Options
 
 ```
-  -h, --help                  help for kyverno-json
-      --payload string        Path to payload (json or yaml file)
-      --policy strings        Path to kyverno-json policies
-      --pre-process strings   JmesPath expression used to pre process payload
+  -h, --help   help for kyverno-json
 ```
 
 ### SEE ALSO
 
 * [kyverno-json completion](kyverno-json_completion.md)	 - Generate the autocompletion script for the specified shell
 * [kyverno-json docs](kyverno-json_docs.md)	 - Generates reference documentation.
+* [kyverno-json scan](kyverno-json_scan.md)	 - scan
