+++ docs/user/commands/kyverno-json_scan.md
@@ -1,0 +1,25 @@
+## kyverno-json scan
+
+scan
+
+### Synopsis
+
+Apply policies to json resources
+
+```
+kyverno-json scan [flags]
+```
+
+### Options
+
+```
+  -h, --help                  help for scan
+      --payload string        Path to payload (json or yaml file)
+      --policy strings        Path to kyverno-json policies
+      --pre-process strings   JmesPath expression used to pre process payload
+```
+
+### SEE ALSO
+
+* [kyverno-json](kyverno-json.md)	 - kyverno-json
+
