+++ pkg/commands/scan/doc.go
@@ -1,0 +1,1 @@
+package scan
